requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic
python-dotenv